requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
gspread==5.1.1
h11==0.12.0
httplib2==0.20.2
idna==3.3
imageio==2.13.5
networkx==2.6.3
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.1.1
#opencv-python==4.5.5.62 s
opencv-python-headless==4.5.5.62
outcome==1.1.0
packaging==21.3
Pillow==9.0.0
playwright==1.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDrive==1.3.1
pyee==8.2.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pyshorteners==1.0.1
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
requests==2.27.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
selenium==4.1.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
telegram==0.0.1
tifffile==2021.11.2
TikTokApi==4.1.0
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
websockets==10.1
wget==3.2
wsproto==1.0.0
